added documentation to all public Tracker methods
