+++ PiwikTracker/Tracker.swift
@@ -1,13 +1,24 @@
 import Foundation
 
+/// The Piwik Tracker is a Swift framework to send analytics to the Piwik server.
+///
+/// ## Basic Usage
+/// * Configure the shared instance as early as possible in your application lifecyle.
+/// * Use the track methods to track your views, events and more.
 final public class Tracker: NSObject {
     
-    let dispatcher: Dispatcher
-    var queue: Queue
-    let siteId: String
+    private let dispatcher: Dispatcher
+    private var queue: Queue
+    internal let siteId: String
     
     internal static var _sharedInstance: Tracker?
     
+    /// Create and Configure a new Tracker
+    ///
+    /// - Parameters:
+    ///   - siteId: The unique site id generated by the server when a new site was created.
+    ///   - queue: The queue to use to store all analytics until it is dispatched to the server.
+    ///   - dispatcher: The dispatcher to use to transmit all analytics to the server.
     required public init(siteId: String, queue: Queue, dispatcher: Dispatcher) {
         self.siteId = siteId
         self.queue = queue
@@ -16,15 +27,31 @@ final public class Tracker: NSObject {
         startDispatchTimer()
     }
     
-    func queue(event: Event) {
+    /// Create and Configure a new Tracker
+    ///
+    /// A volatile memory queue will be used to store the analytics data. All not transmitted data will be lost when the application gets terminated.
+    /// The URLSessionDispatcher will be used to transmit the data to the server.
+    ///
+    /// - Parameters:
+    ///   - siteId: The unique site id generated by the server when a new site was created.
+    ///   - baseURL: The url of the piwik server. This url has to end in `piwik.php`.
+    convenience public init(siteId: String, baseURL: URL) {
+        let queue = MemoryQueue()
+        let dispatcher = URLSessionDispatcher(baseURL: baseURL)
+        self.init(siteId: siteId, queue: queue, dispatcher: dispatcher)
+    }
+    
+    internal func queue(event: Event) {
         queue.enqueue(event: event)
     }
     
     // MARK: dispatching
     
     private let numberOfEventsDispatchedAtOnce = 20
-    public private(set) var isDispatching = false
+    private(set) var isDispatching = false
     
+    
+    /// Manually start the dispatching process.
     public func dispatch() {
         guard !isDispatching else { return }
         guard queue.eventCount > 0 else {
@@ -66,29 +93,46 @@ final public class Tracker: NSObject {
         self.dispatchTimer = Timer.scheduledTimer(timeInterval: dispatchInterval, target: self, selector: #selector(dispatch), userInfo: nil, repeats: false)
     }
     
-    var visitor = Visitor.current()
-    var session = Session.current()
+    internal var visitor = Visitor.current()
+    internal var session = Session.current()
 }
 
 // shared instance
 extension Tracker {
+    
+    /// Returns the shared tracker. Will return nil if the tracker was not properly confured before.
     public static var shared: Tracker? {
         get { return _sharedInstance }
     }
     
+    /// Configures the shared instance.
+    ///
+    /// A volatile memory queue will be used to store the analytics data. All not transmitted data will be lost when the application gets terminated.
+    /// The URLSessionDispatcher will be used to transmit the data to the server.
+    ///
+    /// - Parameters:
+    ///   - siteId: The unique site id generated by the server when a new site was created.
+    ///   - baseURL: The url of the piwik server. This url has to end in `piwik.php`.
     public class func configureSharedInstance(withSiteID siteID: String, baseURL: URL) {
         let queue = MemoryQueue()
         let dispatcher = URLSessionDispatcher(baseURL: baseURL)
         self._sharedInstance = Tracker.init(siteId: siteID, queue: queue, dispatcher: dispatcher)
     }
     
+    /// Configures the shared instance.
+    ///
+    /// - Parameters:
+    /// - Parameters:
+    ///   - siteId: The unique site id generated by the server when a new site was created.
+    ///   - queue: The queue to use to store all analytics until it is dispatched to the server.
+    ///   - dispatcher: The dispatcher to use to transmit all analytics to the server.
     public class func configureSharedInstance(withSiteID siteID: String, queue: Queue = MemoryQueue(), dispatcher: Dispatcher) {
         self._sharedInstance = Tracker(siteId: siteID, queue: queue, dispatcher: dispatcher)
     }
 }
 
 extension Tracker {
-    func event(action: [String]) -> Event {
+    internal func event(action: [String]) -> Event {
         return Event(
             siteId: siteId,
             uuid: NSUUID(),
@@ -103,21 +147,12 @@ extension Tracker {
     }
 }
 
-extension Locale {
-    static var httpAcceptLanguage: String {
-        var components: [String] = []
-        for (index, languageCode) in preferredLanguages.enumerated() {
-            let quality = 1.0 - (Double(index) * 0.1)
-            components.append("\(languageCode);q=\(quality)")
-            if quality <= 0.5 {
-                break
-            }
-        }
-        return components.joined(separator: ",")
-    }
-}
-
 extension Tracker {
+    /// Tracks a screenview.
+    ///
+    /// This method can be used to track hierarchical screen names, e.g. screen/settings/register. Use this to create a hierarchical and logical grouping of screen views in the Piwik web interface.
+    ///
+    /// - Parameter view: An array of hierarchical screen names.
     public func track(view: [String]) {
         queue(event: event(action: view))
     }
